+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/metrics/faas_metrics.py
@@ -13,9 +13,11 @@
 # limitations under the License.
 
 
+from typing import Final
+
 from opentelemetry.metrics import Counter, Histogram, Meter
 
-FAAS_COLDSTARTS = "faas.coldstarts"
+FAAS_COLDSTARTS: Final = "faas.coldstarts"
 """
 Number of invocation cold starts
 Instrument: counter
@@ -26,13 +28,13 @@ Unit: {coldstart}
 def create_faas_coldstarts(meter: Meter) -> Counter:
     """Number of invocation cold starts"""
     return meter.create_counter(
-        name="faas.coldstarts",
+        name=FAAS_COLDSTARTS,
         description="Number of invocation cold starts",
         unit="{coldstart}",
     )
 
 
-FAAS_CPU_USAGE = "faas.cpu_usage"
+FAAS_CPU_USAGE: Final = "faas.cpu_usage"
 """
 Distribution of CPU usage per invocation
 Instrument: histogram
@@ -43,13 +45,13 @@ Unit: s
 def create_faas_cpu_usage(meter: Meter) -> Histogram:
     """Distribution of CPU usage per invocation"""
     return meter.create_histogram(
-        name="faas.cpu_usage",
+        name=FAAS_CPU_USAGE,
         description="Distribution of CPU usage per invocation",
         unit="s",
     )
 
 
-FAAS_ERRORS = "faas.errors"
+FAAS_ERRORS: Final = "faas.errors"
 """
 Number of invocation errors
 Instrument: counter
@@ -60,13 +62,13 @@ Unit: {error}
 def create_faas_errors(meter: Meter) -> Counter:
     """Number of invocation errors"""
     return meter.create_counter(
-        name="faas.errors",
+        name=FAAS_ERRORS,
         description="Number of invocation errors",
         unit="{error}",
     )
 
 
-FAAS_INIT_DURATION = "faas.init_duration"
+FAAS_INIT_DURATION: Final = "faas.init_duration"
 """
 Measures the duration of the function's initialization, such as a cold start
 Instrument: histogram
@@ -77,13 +79,13 @@ Unit: s
 def create_faas_init_duration(meter: Meter) -> Histogram:
     """Measures the duration of the function's initialization, such as a cold start"""
     return meter.create_histogram(
-        name="faas.init_duration",
+        name=FAAS_INIT_DURATION,
         description="Measures the duration of the function's initialization, such as a cold start",
         unit="s",
     )
 
 
-FAAS_INVOCATIONS = "faas.invocations"
+FAAS_INVOCATIONS: Final = "faas.invocations"
 """
 Number of successful invocations
 Instrument: counter
@@ -94,13 +96,13 @@ Unit: {invocation}
 def create_faas_invocations(meter: Meter) -> Counter:
     """Number of successful invocations"""
     return meter.create_counter(
-        name="faas.invocations",
+        name=FAAS_INVOCATIONS,
         description="Number of successful invocations",
         unit="{invocation}",
     )
 
 
-FAAS_INVOKE_DURATION = "faas.invoke_duration"
+FAAS_INVOKE_DURATION: Final = "faas.invoke_duration"
 """
 Measures the duration of the function's logic execution
 Instrument: histogram
@@ -111,13 +113,13 @@ Unit: s
 def create_faas_invoke_duration(meter: Meter) -> Histogram:
     """Measures the duration of the function's logic execution"""
     return meter.create_histogram(
-        name="faas.invoke_duration",
+        name=FAAS_INVOKE_DURATION,
         description="Measures the duration of the function's logic execution",
         unit="s",
     )
 
 
-FAAS_MEM_USAGE = "faas.mem_usage"
+FAAS_MEM_USAGE: Final = "faas.mem_usage"
 """
 Distribution of max memory usage per invocation
 Instrument: histogram
@@ -128,13 +130,13 @@ Unit: By
 def create_faas_mem_usage(meter: Meter) -> Histogram:
     """Distribution of max memory usage per invocation"""
     return meter.create_histogram(
-        name="faas.mem_usage",
+        name=FAAS_MEM_USAGE,
         description="Distribution of max memory usage per invocation",
         unit="By",
     )
 
 
-FAAS_NET_IO = "faas.net_io"
+FAAS_NET_IO: Final = "faas.net_io"
 """
 Distribution of net I/O usage per invocation
 Instrument: histogram
@@ -145,13 +147,13 @@ Unit: By
 def create_faas_net_io(meter: Meter) -> Histogram:
     """Distribution of net I/O usage per invocation"""
     return meter.create_histogram(
-        name="faas.net_io",
+        name=FAAS_NET_IO,
         description="Distribution of net I/O usage per invocation",
         unit="By",
     )
 
 
-FAAS_TIMEOUTS = "faas.timeouts"
+FAAS_TIMEOUTS: Final = "faas.timeouts"
 """
 Number of invocation timeouts
 Instrument: counter
@@ -162,7 +164,7 @@ Unit: {timeout}
 def create_faas_timeouts(meter: Meter) -> Counter:
     """Number of invocation timeouts"""
     return meter.create_counter(
-        name="faas.timeouts",
+        name=FAAS_TIMEOUTS,
         description="Number of invocation timeouts",
         unit="{timeout}",
     )