+++ CHANGELOG.md
@@ -9,8 +9,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Do not execute Flask Tests in debug mode
   ([#3956](https://github.com/open-telemetry/opentelemetry-python/pull/3956))
+- Log a warning when a `LogRecord` in `sdk/log` has dropped attributes
+  due to reaching limits
+  ([#3946](https://github.com/open-telemetry/opentelemetry-python/pull/3946))
+- Fix RandomIdGenerator can generate invalid Span/Trace Ids
+  ([#3949](https://github.com/open-telemetry/opentelemetry-python/pull/3949))
 - Add Python 3.12 to tox
   ([#3616](https://github.com/open-telemetry/opentelemetry-python/pull/3616))
+- Improve resource field structure for LogRecords
+  ([#3972](https://github.com/open-telemetry/opentelemetry-python/pull/3972))
+- Update Semantic Conventions code generation scripts:
+  - fix namespace exclusion that resulted in dropping  `os` and `net` namespaces.
+  - add `Final` decorator to constants to prevent collisions
+  - enable mypy and fix detected issues
+  - allow to drop specific attributes in preparation for Semantic Conventions v1.26.0
+  ([#3973](https://github.com/open-telemetry/opentelemetry-python/pull/3966))
+- Update semantic conventions to version 1.26.0.
+  ([#3964](https://github.com/open-telemetry/opentelemetry-python/pull/3964))
+- Use semconv exception attributes for record exceptions in spans
+  ([#3979](https://github.com/open-telemetry/opentelemetry-python/pull/3979))
+- Fix _encode_events assumes events.attributes.dropped exists
+  ([#3965](https://github.com/open-telemetry/opentelemetry-python/pull/3965))
+- Validate links at span creation
+  ([#3991](https://github.com/open-telemetry/opentelemetry-python/pull/3991))
 
 ## Version 1.25.0/0.46b0 (2024-05-30)
 
@@ -62,7 +83,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - units with slash are converted e.g. `m/s` -> `meters_per_second`.
   - The exporter's API is not changed
 - Add parameters for Distros and configurators to configure autoinstrumentation in addition to existing environment variables.
-  ([#3864] (https://github.com/open-telemetry/opentelemetry-python/pull/3864))
+  ([#3864](https://github.com/open-telemetry/opentelemetry-python/pull/3864))
 
 ## Version 1.24.0/0.45b0 (2024-03-28)